refactor - copilot-provider files
