+++ CopilotKit/packages/react-core/src/components/copilot-provider/index.ts
@@ -1,0 +1,10 @@
+export {
+  CopilotProvider,
+  defaultCopilotContextCategories,
+} from "./copilot-provider";
+
+export type {
+  CopilotProviderApiEndpointProps,
+  CopilotProviderApiConfigProps,
+  CopilotProviderProps,
+} from "./copilot-provider-props";
